requirements.txt
cython>=0.19
dash>=1.6.1
-e git://github.com/plotly/dash-bio.git#egg=dash_bio
dash-bio-utils==0.0.4
dash-daq==0.2.2
gunicorn==19.9.0
jsonschema==2.6.0
matplotlib==3.0.2
numpy~=1.21.3
pandas>=0.24.2
plotly>=3.5.0
PubChemPy==1.0.4
requests==2.21.0
scikit-learn==0.20.2
scipy>=1.1.0

PyYAML~=6.0
pyodbc~=4.0.32
SQLAlchemy~=1.4.26
xlwings~=0.24.9
tabula~=1.0.5